refactor: decompose bridge.ts into four focused modules
Extract the monolithic AlgoChatBridge (~1856 lines) into four independently
testable services with clear responsibility boundaries:

- ResponseFormatter (304 lines): On-chain message sending, PSK chunking,
  ALGO spend tracking, event emission, and public key caching
- CommandHandler (428 lines): Slash command parsing, authorization checks,
  and dispatch (/status, /stop, /agent, /credits, /work, /council, etc.)
- SubscriptionManager (647 lines): Session event subscriptions for both
  on-chain and local (browser) responses, progress tracking, ack delays,
  and timeout management
- DiscoveryService (244 lines): Sender discovery via indexer, conversation
  seeding, agent wallet caching, fast-polling for approvals, and default
  agent/project resolution

bridge.ts is now a thin orchestration layer (694 lines, 63% reduction)
that composes these services and handles incoming message routing.

Architectural decisions:
- All public API surface preserved (re-exports types from extracted modules)
- Callers (index.ts, ws/handler.ts, routes/index.ts) require zero changes
- Follows existing codebase patterns for service injection (constructor +
  setter methods for optional/circular dependencies)
- JSDoc on all public interfaces
- Lambda-based dependency injection between services avoids circular refs

Validation: TypeScript compiles clean, all 122 tests pass.
