+++ server/algochat/bridge.ts
@@ -1,3 +1,17 @@
+/**
+ * AlgoChatBridge — Thin orchestration layer composing four focused services:
+ *
+ * - **ResponseFormatter** — Message sending, on-chain delivery, event emission
+ * - **CommandHandler** — Slash command parsing, authorization, dispatch
+ * - **SubscriptionManager** — Session event subscriptions and response lifecycle
+ * - **DiscoveryService** — Agent/sender discovery, conversation seeding, polling
+ *
+ * This module wires the services together and handles incoming message routing.
+ * All business logic lives in the extracted modules; this file is purely
+ * orchestration and lifecycle management.
+ *
+ * @module
+ */
 import type { Database } from 'bun:sqlite';
 import type { ProcessManager } from '../process/manager';
 import type { AlgoChatConfig } from './config';
@@ -9,10 +23,9 @@ import {
     updateConversationRound,
     listConversations,
 } from '../db/sessions';
-import { getAgent, getAlgochatEnabledAgents, listAgents } from '../db/agents';
+import { getAgent } from '../db/agents';
 import { createSession } from '../db/sessions';
 import type { ClaudeStreamEvent } from '../process/types';
-import { extractContentText } from '../process/types';
 import { PSKManager } from './psk';
 import type { AgentWalletService } from './agent-wallet';
 import type { AgentDirectory } from './agent-directory';
@@ -21,77 +34,64 @@ import type { ApprovalManager } from '../process/approval-manager';
 import type { ApprovalRequestWire } from '../process/approval-types';
 import type { WorkTaskService } from '../work/service';
 import { formatApprovalForChain, parseApprovalResponse } from './approval-format';
-import { checkAlgoLimit, recordAlgoSpend } from '../db/spending';
 import {
     getBalance,
     purchaseCredits,
     maybeGrantFirstTimeCredits,
     canStartSession,
     getCreditConfig,
-    getTransactionHistory,
 } from '../db/credits';
-import { updateSessionAlgoSpent } from '../db/sessions';
 import { parseGroupPrefix, reassembleGroupMessage } from './group-sender';
-import { saveAlgoChatMessage } from '../db/algochat-messages';
 import { createLogger } from '../lib/logger';
-import { listCouncils, createCouncil, getCouncilLaunch } from '../db/councils';
-import { launchCouncil, onCouncilStageChange } from '../routes/councils';
+
+// Composed services
+import { ResponseFormatter } from './response-formatter';
+import { CommandHandler } from './command-handler';
+import { SubscriptionManager } from './subscription-manager';
+import { DiscoveryService } from './discovery-service';
+
+// Re-export types from extracted modules so callers don't need to change imports
+export type { AlgoChatEventCallback } from './response-formatter';
+export type { LocalChatSendFn, LocalChatEvent, LocalChatEventFn } from './subscription-manager';
 
 const log = createLogger('AlgoChatBridge');
 
-const SUBSCRIPTION_TIMEOUT_MS = 10 * 60 * 1000; // 10 minutes (activity resets timer)
-const PUBLIC_KEY_CACHE_TTL_MS = 60 * 60 * 1000; // 1 hour
-
-export type AlgoChatEventCallback = (
-    participant: string,
-    content: string,
-    direction: 'inbound' | 'outbound' | 'status',
-    fee?: number,
-) => void;
-
-export type LocalChatSendFn = (
-    participant: string,
-    content: string,
-    direction: 'inbound' | 'outbound',
-) => void;
-
-export type LocalChatEvent =
-    | { type: 'message'; content: string; direction: 'inbound' | 'outbound' }
-    | { type: 'stream'; chunk: string; done: boolean }
-    | { type: 'tool_use'; toolName: string; input: string }
-    | { type: 'thinking'; active: boolean }
-    | { type: 'session_info'; sessionId: string };
-
-export type LocalChatEventFn = (event: LocalChatEvent) => void;
-
-interface CachedPublicKey {
-    key: Uint8Array;
-    cachedAt: number;
-}
-
+/**
+ * Central orchestrator for the AlgoChat system.
+ *
+ * Bridges on-chain Algorand messaging with the agent session system.
+ * Composes four focused services and handles message routing between them.
+ *
+ * Public API surface is preserved for backward compatibility — callers
+ * (server/index.ts, ws/handler.ts, routes/index.ts) require no changes.
+ */
 export class AlgoChatBridge {
     private db: Database;
     private processManager: ProcessManager;
     private config: AlgoChatConfig;
     private service: AlgoChatService;
+
+    // Composed services
+    private responseFormatter: ResponseFormatter;
+    private commandHandler: CommandHandler;
+    private subscriptionManager: SubscriptionManager;
+    private discoveryService: DiscoveryService;
+
+    // Optional dependencies
     private agentWalletService: AgentWalletService | null = null;
     private agentDirectory: AgentDirectory | null = null;
-    private eventCallbacks: Set<AlgoChatEventCallback> = new Set();
-    private publicKeyCache: Map<string, CachedPublicKey> = new Map();
-    private localAgentSessions: Map<string, string> = new Map();
-    private localSubscriptions: Map<string, (sid: string, event: ClaudeStreamEvent) => void> = new Map();
-    private localSendFns: Map<string, LocalChatSendFn> = new Map();
-    private localEventFns: Map<string, LocalChatEventFn> = new Map();
-    private chainSubscriptions: Set<string> = new Set();
-    private processedTxids: Set<string> = new Set();
-    private subscriptionTimers: Map<string, ReturnType<typeof setTimeout>> = new Map();
-    private pskManager: PSKManager | null = null;
     private approvalManager: ApprovalManager | null = null;
-    private workTaskService: WorkTaskService | null = null;
-    private agentMessengerRef: AgentMessenger | null = null;
-    private fastPollTimer: ReturnType<typeof setInterval> | null = null;
-    private discoveryTimer: ReturnType<typeof setInterval> | null = null;
+    private pskManager: PSKManager | null = null;
     private sessionNotificationCallback: ((sid: string, event: ClaudeStreamEvent) => void) | null = null;
+
+    // Local chat state (kept here as it bridges WS handler → subscription manager)
+    private localAgentSessions: Map<string, string> = new Map();
+
+    // On-chain message dedup
+    private processedTxids: Set<string> = new Set();
+
+    // Group message reassembly buffer
+    private pendingGroupChunks: Map<string, { chunks: unknown[]; firstSeen: number }> = new Map();
 
     constructor(
         db: Database,
@@ -104,34 +104,58 @@ export class AlgoChatBridge {
         this.config = config;
         this.service = service;
 
+        // Initialize composed services.
+        // Note: isOwnerFn uses a lambda that closes over `this` so it's safe
+        // to pass before commandHandler is assigned — the lambda is only
+        // called later at runtime, never during construction.
+        this.responseFormatter = new ResponseFormatter(db, config, service);
+        this.discoveryService = new DiscoveryService(db, config, service, (p) => this.commandHandler.isOwner(p));
+        this.subscriptionManager = new SubscriptionManager(processManager, this.responseFormatter);
+        this.commandHandler = new CommandHandler(db, config, processManager, this.responseFormatter, {
+            findAgentForNewConversation: () => this.discoveryService.findAgentForNewConversation(),
+            getDefaultProjectId: () => this.discoveryService.getDefaultProjectId(),
+        });
+
         this.setupMessageHandler();
         this.setupPSKManager();
         this.setupSessionNotifications();
     }
 
+    // ── Dependency injection (preserves existing API) ────────────────────
+
+    /** Inject the agent wallet service for per-agent on-chain sends. */
     setAgentWalletService(service: AgentWalletService): void {
         this.agentWalletService = service;
+        this.responseFormatter.setAgentWalletService(service);
     }
 
+    /** Get the agent wallet service (used by WS handler for agent_reward). */
     getAgentWalletService(): AgentWalletService | null {
         return this.agentWalletService;
     }
 
+    /** Inject the agent directory for agent-to-agent message filtering. */
     setAgentDirectory(directory: AgentDirectory): void {
         this.agentDirectory = directory;
     }
 
+    /** Inject the approval manager for handling tool approval requests. */
     setApprovalManager(manager: ApprovalManager): void {
         this.approvalManager = manager;
+        this.discoveryService.setApprovalManager(manager);
     }
 
+    /** Inject the work task service for /work command support. */
     setWorkTaskService(service: WorkTaskService): void {
-        this.workTaskService = service;
+        this.commandHandler.setWorkTaskService(service);
     }
 
+    /** Inject the agent messenger for council and inter-agent messaging. */
     setAgentMessenger(messenger: AgentMessenger): void {
-        this.agentMessengerRef = messenger;
+        this.commandHandler.setAgentMessenger(messenger);
     }
+
+    // ── Public API ───────────────────────────────────────────────────────
 
     /**
      * Send an approval request to a participant on-chain.
@@ -143,45 +167,45 @@ export class AlgoChatBridge {
             toolInput: {},
             source: 'algochat',
         });
-        await this.sendResponse(participant, formatted);
-        this.startFastPolling();
+        await this.responseFormatter.sendResponse(participant, formatted);
+        this.discoveryService.startFastPolling();
     }
 
+    /** Start all AlgoChat services. */
     start(): void {
-        this.seedConversations();
+        this.discoveryService.seedConversations();
         this.service.syncManager.start();
         this.pskManager?.start(this.config.syncInterval);
-        this.startDiscoveryPolling();
+        this.discoveryService.startDiscoveryPolling();
         log.info('Started listening for messages');
     }
 
+    /** Stop all AlgoChat services and clean up resources. */
     stop(): void {
         this.service.syncManager.stop();
         this.pskManager?.stop();
-        this.stopFastPolling();
-        this.stopDiscoveryPolling();
+        this.discoveryService.cleanup();
         // Unsubscribe session notification handler
         if (this.sessionNotificationCallback) {
             this.processManager.unsubscribeAll(this.sessionNotificationCallback);
             this.sessionNotificationCallback = null;
         }
-        // Clear all subscription timers and chain subscriptions
-        for (const timer of this.subscriptionTimers.values()) {
-            clearTimeout(timer);
-        }
-        this.subscriptionTimers.clear();
-        this.chainSubscriptions.clear();
+        // Clean up subscription timers
+        this.subscriptionManager.cleanup();
         log.info('Stopped');
     }
 
-    onEvent(callback: AlgoChatEventCallback): void {
-        this.eventCallbacks.add(callback);
+    /** Register a callback for AlgoChat feed events. */
+    onEvent(callback: import('./response-formatter').AlgoChatEventCallback): void {
+        this.responseFormatter.onEvent(callback);
     }
 
-    offEvent(callback: AlgoChatEventCallback): void {
-        this.eventCallbacks.delete(callback);
+    /** Unregister a feed event callback. */
+    offEvent(callback: import('./response-formatter').AlgoChatEventCallback): void {
+        this.responseFormatter.offEvent(callback);
     }
 
+    /** Get the current AlgoChat status. */
     getStatus(): AlgoChatStatus {
         const conversations = listConversations(this.db);
         return {
@@ -195,15 +219,15 @@ export class AlgoChatBridge {
 
     /**
      * Handle a message from the browser dashboard chat UI.
-     * Routes through the same agent->session->process flow, but sends the
+     * Routes through the same agent→session→process flow, but sends the
      * response back via the provided callback instead of on-chain.
      */
     async handleLocalMessage(
         agentId: string,
         content: string,
-        sendFn: LocalChatSendFn,
+        sendFn: import('./subscription-manager').LocalChatSendFn,
         projectId?: string,
-        eventFn?: LocalChatEventFn,
+        eventFn?: import('./subscription-manager').LocalChatEventFn,
     ): Promise<void> {
         log.debug('handleLocalMessage', { agentId, content: content.slice(0, 50) });
         const agent = getAgent(this.db, agentId);
@@ -229,9 +253,9 @@ export class AlgoChatBridge {
         // Update the sendFn so responses go to the current WS connection
         const existingSessionId = this.localAgentSessions.get(agentId);
         if (existingSessionId) {
-            this.localSendFns.set(existingSessionId, sendFn);
+            this.subscriptionManager.updateLocalSendFn(existingSessionId, sendFn);
             if (eventFn) {
-                this.localEventFns.set(existingSessionId, eventFn);
+                this.subscriptionManager.updateLocalEventFn(existingSessionId, eventFn);
             }
         }
 
@@ -241,20 +265,18 @@ export class AlgoChatBridge {
             if (sent) {
                 log.debug(`Sent message to running session ${existingSessionId}`);
                 eventFn?.({ type: 'session_info', sessionId: existingSessionId });
-                this.subscribeForLocalResponse(existingSessionId, sendFn);
+                this.subscriptionManager.subscribeForLocalResponse(existingSessionId, sendFn);
                 return;
             }
 
             // Process not running — clear stale entry and create a fresh session below
             log.debug(`Stale session ${existingSessionId}, creating new one`);
             this.localAgentSessions.delete(agentId);
-            this.localSubscriptions.delete(existingSessionId);
-            this.localSendFns.delete(existingSessionId);
-            this.localEventFns.delete(existingSessionId);
+            this.subscriptionManager.cleanupLocalSession(existingSessionId);
         }
 
         // Create a new session
-        const resolvedProjectId = projectId ?? agent.defaultProjectId ?? this.getDefaultProjectId();
+        const resolvedProjectId = projectId ?? agent.defaultProjectId ?? this.discoveryService.getDefaultProjectId();
         log.debug(`Creating new session`, { projectId: resolvedProjectId, agentId });
         const session = createSession(this.db, {
             projectId: resolvedProjectId,
@@ -266,357 +288,16 @@ export class AlgoChatBridge {
 
         log.debug(`Session created: ${session.id}, starting process`);
         this.localAgentSessions.set(agentId, session.id);
-        this.localSendFns.set(session.id, sendFn);
+        this.subscriptionManager.updateLocalSendFn(session.id, sendFn);
         if (eventFn) {
-            this.localEventFns.set(session.id, eventFn);
+            this.subscriptionManager.updateLocalEventFn(session.id, eventFn);
         }
         eventFn?.({ type: 'session_info', sessionId: session.id });
-        this.subscribeForLocalResponse(session.id, sendFn);
+        this.subscriptionManager.subscribeForLocalResponse(session.id, sendFn);
         this.processManager.startProcess(session, content);
     }
 
-    /** Check if a participant is authorized to run privileged commands. Fail-closed: returns false when no owners configured. */
-    private isOwner(participant: string): boolean {
-        if (this.config.ownerAddresses.size === 0) {
-            log.warn('Owner check failed — no owner addresses configured', { participant: participant.slice(0, 8) });
-            return false;
-        }
-        if (!this.config.ownerAddresses.has(participant)) {
-            log.debug('Non-owner address', { participant: participant.slice(0, 8) });
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * Handle commands from AlgoChat messages.
-     * Returns true if the message was handled as a command.
-     */
-    private handleCommand(participant: string, content: string): boolean {
-        const trimmed = content.trim();
-        if (!trimmed.startsWith('/')) return false;
-
-        const parts = trimmed.split(/\s+/);
-        const command = parts[0].toLowerCase();
-
-        // Privileged commands require owner authorization
-        const PRIVILEGED_COMMANDS = new Set(['/stop', '/approve', '/deny', '/mode', '/work', '/agent', '/council']);
-        if (PRIVILEGED_COMMANDS.has(command) && !this.isOwner(participant)) {
-            log.warn('Unauthorized command attempt', { participant: participant.slice(0, 8), command });
-            this.sendResponse(participant, `Unauthorized: ${command} requires owner access`);
-            return true;
-        }
-
-        switch (command) {
-            case '/status': {
-                const activeCount = this.processManager.getActiveSessionIds().length;
-                const conversations = listConversations(this.db);
-                this.sendResponse(participant, `Active sessions: ${activeCount}, conversations: ${conversations.length}`);
-                return true;
-            }
-
-            case '/stop': {
-                const sessionId = parts[1];
-                if (!sessionId) {
-                    this.sendResponse(participant, 'Usage: /stop <session-id>');
-                    return true;
-                }
-                if (this.processManager.isRunning(sessionId)) {
-                    this.processManager.stopProcess(sessionId);
-                    this.sendResponse(participant, `Stopped session ${sessionId}`);
-                } else {
-                    this.sendResponse(participant, `Session ${sessionId} is not running`);
-                }
-                return true;
-            }
-
-            case '/agent': {
-                const agentName = parts.slice(1).join(' ');
-                if (!agentName) {
-                    const agents = getAlgochatEnabledAgents(this.db);
-                    const names = agents.map((a) => a.name).join(', ');
-                    this.sendResponse(participant, `Available agents: ${names || 'none'}`);
-                    return true;
-                }
-                // Route subsequent messages to the specified agent
-                const agents = getAlgochatEnabledAgents(this.db);
-                const matched = agents.find((a) => a.name.toLowerCase() === agentName.toLowerCase());
-                if (matched) {
-                    this.config.defaultAgentId = matched.id;
-                    this.sendResponse(participant, `Routing to agent: ${matched.name}`);
-                } else {
-                    this.sendResponse(participant, `Agent "${agentName}" not found`);
-                }
-                return true;
-            }
-
-            case '/queue': {
-                const queued = this.processManager.approvalManager.getQueuedRequests();
-                if (queued.length === 0) {
-                    this.sendResponse(participant, 'No pending escalation requests');
-                } else {
-                    const lines = queued.map((q) => `#${q.id}: [${q.toolName}] session=${q.sessionId.slice(0, 8)} (${q.createdAt})`);
-                    this.sendResponse(participant, `Pending escalations:\n${lines.join('\n')}`);
-                }
-                return true;
-            }
-
-            case '/approve': {
-                const queueId = parseInt(parts[1], 10);
-                if (isNaN(queueId)) {
-                    this.sendResponse(participant, 'Usage: /approve <queue-id>');
-                    return true;
-                }
-                const resolved = this.processManager.approvalManager.resolveQueuedRequest(queueId, true);
-                this.sendResponse(participant, resolved
-                    ? `Escalation #${queueId} approved`
-                    : `Escalation #${queueId} not found or already resolved`);
-                return true;
-            }
-
-            case '/deny': {
-                const queueId = parseInt(parts[1], 10);
-                if (isNaN(queueId)) {
-                    this.sendResponse(participant, 'Usage: /deny <queue-id>');
-                    return true;
-                }
-                const resolved = this.processManager.approvalManager.resolveQueuedRequest(queueId, false);
-                this.sendResponse(participant, resolved
-                    ? `Escalation #${queueId} denied`
-                    : `Escalation #${queueId} not found or already resolved`);
-                return true;
-            }
-
-            case '/mode': {
-                const newMode = parts[1]?.toLowerCase();
-                if (!newMode) {
-                    this.sendResponse(participant, `Current mode: ${this.processManager.approvalManager.operationalMode}`);
-                    return true;
-                }
-                const validModes = ['normal', 'queued', 'paused'];
-                if (!validModes.includes(newMode)) {
-                    this.sendResponse(participant, `Invalid mode. Use: ${validModes.join(', ')}`);
-                    return true;
-                }
-                this.processManager.approvalManager.operationalMode = newMode as 'normal' | 'queued' | 'paused';
-                this.sendResponse(participant, `Mode set to: ${newMode}`);
-                return true;
-            }
-
-            case '/credits': {
-                const balance = getBalance(this.db, participant);
-                const config = getCreditConfig(this.db);
-                const lines = [
-                    `💰 Credit Balance:`,
-                    `  Available: ${balance.available} credits`,
-                    `  Reserved: ${balance.reserved} credits`,
-                    `  Total: ${balance.credits} credits`,
-                    `  Purchased: ${balance.totalPurchased} | Used: ${balance.totalConsumed}`,
-                    ``,
-                    `📊 Rates:`,
-                    `  1 ALGO = ${config.creditsPerAlgo} credits`,
-                    `  1 turn = ${config.creditsPerTurn} credit(s)`,
-                    `  1 agent message = ${config.creditsPerAgentMessage} credit(s)`,
-                    ``,
-                    `Send ALGO to this address to purchase credits.`,
-                ];
-                this.sendResponse(participant, lines.join('\n'));
-                return true;
-            }
-
-            case '/history': {
-                const limit = parseInt(parts[1], 10) || 10;
-                const transactions = getTransactionHistory(this.db, participant, Math.min(limit, 20));
-                if (transactions.length === 0) {
-                    this.sendResponse(participant, 'No credit transactions yet.');
-                    return true;
-                }
-                const lines = transactions.map((t) =>
-                    `${t.type === 'purchase' || t.type === 'grant' ? '+' : '-'}${t.amount} [${t.type}] → bal:${t.balanceAfter} (${t.createdAt})`
-                );
-                this.sendResponse(participant, `📜 Recent Transactions:\n${lines.join('\n')}`);
-                return true;
-            }
-
-            case '/work': {
-                const description = parts.slice(1).join(' ');
-                if (!description) {
-                    this.sendResponse(participant, 'Usage: /work <task description>');
-                    return true;
-                }
-
-                if (!this.workTaskService) {
-                    this.sendResponse(participant, 'Work task service not available');
-                    return true;
-                }
-
-                const agentId = this.findAgentForNewConversation();
-                if (!agentId) {
-                    this.sendResponse(participant, 'No agent available for work tasks');
-                    return true;
-                }
-
-                this.workTaskService.create({
-                    agentId,
-                    description,
-                    source: 'algochat',
-                    requesterInfo: { participant },
-                }).then((task) => {
-                    this.sendResponse(participant, `Work task started: ${task.id}\nBranch: ${task.branchName ?? 'creating...'}\nStatus: ${task.status}`);
-
-                    this.workTaskService?.onComplete(task.id, (completed) => {
-                        if (completed.status === 'completed' && completed.prUrl) {
-                            this.sendResponse(participant, `Work task completed!\nPR: ${completed.prUrl}`);
-                        } else {
-                            this.sendResponse(participant, `Work task failed: ${completed.error ?? 'Unknown error'}`);
-                        }
-                    });
-                }).catch((err) => {
-                    this.sendResponse(participant, `Work task error: ${err instanceof Error ? err.message : String(err)}`);
-                });
-                return true;
-            }
-
-            case '/council': {
-                this.handleCouncilCommand(participant, parts).catch((err) => {
-                    this.sendResponse(participant, `Council error: ${err instanceof Error ? err.message : String(err)}`);
-                });
-                return true;
-            }
-
-            default:
-                return false;
-        }
-    }
-
-    /**
-     * Handle the `/council` command from AlgoChat.
-     *
-     * Usage:
-     *   /council <prompt>                       — auto-create council with all enabled agents
-     *   /council MyCouncilName -- <prompt>      — use existing council by name
-     */
-    private async handleCouncilCommand(participant: string, parts: string[]): Promise<void> {
-        const rest = parts.slice(1).join(' ').trim();
-        if (!rest) {
-            this.sendResponse(participant, 'Usage:\n  /council <prompt>\n  /council <CouncilName> -- <prompt>');
-            return;
-        }
-
-        // Parse: "/council CouncilName -- prompt" or "/council prompt"
-        const doubleDashIdx = rest.indexOf('--');
-        let councilName: string | null = null;
-        let prompt: string;
-
-        if (doubleDashIdx >= 0) {
-            councilName = rest.slice(0, doubleDashIdx).trim();
-            prompt = rest.slice(doubleDashIdx + 2).trim();
-        } else {
-            prompt = rest;
-        }
-
-        if (!prompt) {
-            this.sendResponse(participant, 'Please provide a prompt for the council.');
-            return;
-        }
-
-        // Resolve or auto-create the council
-        let councilId: string;
-        let councilLabel: string;
-
-        if (councilName) {
-            // Find existing council by name
-            const councils = listCouncils(this.db);
-            const match = councils.find((c) => c.name.toLowerCase() === councilName!.toLowerCase());
-            if (!match) {
-                const available = councils.map((c) => c.name).join(', ');
-                this.sendResponse(participant, `Council "${councilName}" not found.\nAvailable: ${available || 'none'}`);
-                return;
-            }
-            councilId = match.id;
-            councilLabel = match.name;
-        } else {
-            // Auto-create council with all algochat-enabled agents
-            const agents = getAlgochatEnabledAgents(this.db);
-            if (agents.length === 0) {
-                this.sendResponse(participant, 'No AlgoChat-enabled agents available for council.');
-                return;
-            }
-            const agentIds = agents.map((a) => a.id);
-            const chairmanId = agents[0].id; // First agent becomes chairman
-            const council = createCouncil(this.db, {
-                name: `AlgoChat Council ${new Date().toISOString().slice(0, 16)}`,
-                description: 'Auto-created from AlgoChat /council command',
-                agentIds,
-                chairmanAgentId: chairmanId,
-                discussionRounds: 2,
-            });
-            councilId = council.id;
-            councilLabel = council.name;
-        }
-
-        // Resolve project ID (reuse existing helper)
-        const projectId = this.getDefaultProjectId();
-
-        // Launch the council
-        this.sendResponse(participant, `Launching council "${councilLabel}"...\nPrompt: ${prompt.slice(0, 200)}`);
-
-        try {
-            const result = launchCouncil(
-                this.db,
-                this.processManager,
-                councilId,
-                projectId,
-                prompt,
-                this.agentMessengerRef,
-            );
-
-            this.sendResponse(participant, `Council launched! (${result.sessionIds.length} agents responding)\nLaunch ID: ${result.launchId.slice(0, 8)}...`);
-
-            // Monitor stage changes and relay progress + final synthesis on-chain.
-            // Safety timeout prevents the listener from leaking if the council
-            // pipeline crashes before reaching the 'complete' stage.
-            const COUNCIL_LISTENER_TIMEOUT_MS = 45 * 60 * 1000; // 45 minutes
-
-            const cleanup = () => {
-                clearTimeout(safetyTimer);
-                unsubscribe();
-            };
-
-            const safetyTimer = setTimeout(() => {
-                log.warn('Council stage listener timed out, cleaning up', { launchId: result.launchId });
-                unsubscribe();
-            }, COUNCIL_LISTENER_TIMEOUT_MS);
-
-            const unsubscribe = onCouncilStageChange((launchId, stage) => {
-                if (launchId !== result.launchId) return;
-
-                if (stage === 'discussing') {
-                    this.sendResponse(participant, `[Council] Agents are now discussing...`);
-                } else if (stage === 'reviewing') {
-                    this.sendResponse(participant, `[Council] Peer review stage started.`);
-                } else if (stage === 'synthesizing') {
-                    this.sendResponse(participant, `[Council] Chairman is synthesizing final answer...`);
-                } else if (stage === 'complete') {
-                    cleanup();
-                    // Fetch the synthesis and send it back
-                    const launch = getCouncilLaunch(this.db, result.launchId);
-                    if (launch?.synthesis) {
-                        const MAX_SYNTHESIS_LENGTH = 3000;
-                        const synthesis = launch.synthesis.length > MAX_SYNTHESIS_LENGTH
-                            ? launch.synthesis.slice(0, MAX_SYNTHESIS_LENGTH) + '\n\n[Truncated — view full synthesis on dashboard]'
-                            : launch.synthesis;
-                        this.sendResponse(participant, `[Council Complete]\n\n${synthesis}`);
-                    } else {
-                        this.sendResponse(participant, `[Council Complete] No synthesis produced.`);
-                    }
-                }
-            });
-        } catch (err) {
-            const msg = err instanceof Error ? err.message : String(err);
-            this.sendResponse(participant, `Council launch failed: ${msg}`);
-        }
-    }
+    // ── Message handling (private orchestration) ─────────────────────────
 
     private setupMessageHandler(): void {
         this.service.syncManager.on('onMessagesReceived', (participant, messages) => {
@@ -625,18 +306,16 @@ export class AlgoChatBridge {
             const regularMessages: typeof messages = [];
 
             // Collect known agent wallet addresses to filter outbound messages
-            // sent from per-agent wallets (which the sync sees as 'received')
-            const agentWalletAddresses = this.getAgentWalletAddresses();
+            const agentWalletAddresses = this.discoveryService.getAgentWalletAddresses();
 
             for (const msg of messages) {
                 if (msg.direction === 'sent') continue;
 
-                // Skip messages sent by our agent wallets (sync sees them as
-                // 'received' because the sender doesn't match the main account)
+                // Skip messages sent by our agent wallets
                 const sender = (msg as unknown as { sender?: string }).sender;
                 if (sender && agentWalletAddresses.has(sender)) continue;
 
-                // Dedup by transaction ID — skip messages we've already processed
+                // Dedup by transaction ID
                 const txid = (msg as unknown as { id?: string }).id;
                 if (txid) {
                     if (this.processedTxids.has(txid)) {
@@ -656,9 +335,7 @@ export class AlgoChatBridge {
                 }
             }
 
-            // Prune old txids to prevent unbounded growth (keep last 500).
-            // JS Sets iterate in insertion order, so dropping from the front
-            // removes the oldest entries.
+            // Prune old txids to prevent unbounded growth (keep last 500)
             if (this.processedTxids.size > 500) {
                 const all = [...this.processedTxids];
                 this.processedTxids = new Set(all.slice(all.length - 500));
@@ -694,9 +371,6 @@ export class AlgoChatBridge {
             }
         });
     }
-
-    /** Buffer for incomplete group chunks that span multiple sync batches. */
-    private pendingGroupChunks: Map<string, { chunks: unknown[]; firstSeen: number }> = new Map();
 
     private bufferGroupChunk(participant: string, msg: unknown): void {
         const message = msg as { content: string; confirmedRound: number | bigint; fee?: number };
@@ -738,6 +412,7 @@ export class AlgoChatBridge {
         if (!this.config.pskContact) return;
 
         this.pskManager = new PSKManager(this.db, this.service, this.config.pskContact, this.config.network);
+        this.responseFormatter.setPskManager(this.pskManager);
 
         this.pskManager.onMessage((msg) => {
             this.handleIncomingMessage(msg.sender, msg.content, msg.confirmedRound).catch((err) => {
@@ -799,13 +474,21 @@ export class AlgoChatBridge {
                     participant: conversation.participantAddr,
                     error: event.error.message,
                 });
-                this.sendResponse(conversation.participantAddr, `[Error: ${event.error.message}]`);
+                this.responseFormatter.sendResponse(conversation.participantAddr, `[Error: ${event.error.message}]`);
             }
         };
         this.sessionNotificationCallback = callback;
         this.processManager.subscribeAll(callback);
     }
 
+    /**
+     * Core incoming message handler — routes messages through the pipeline:
+     * 1. Safety guards (raw group chunks, approval responses, agent-to-agent)
+     * 2. Owner authorization
+     * 3. Credit system (payments, first-time grants)
+     * 4. Command dispatch (via CommandHandler)
+     * 5. Session creation/resumption (via SubscriptionManager)
+     */
     private async handleIncomingMessage(
         participant: string,
         content: string,
@@ -834,7 +517,7 @@ export class AlgoChatBridge {
                         shortId: approvalResponse.shortId,
                         behavior: approvalResponse.behavior,
                     });
-                    this.stopFastPolling();
+                    this.discoveryService.stopFastPolling();
                     return;
                 }
             }
@@ -850,21 +533,18 @@ export class AlgoChatBridge {
         }
 
         // Owner-only mode: only respond to wallets in ALGOCHAT_OWNER_ADDRESSES
-        // This is the sole access control gate — the database allowlist is not used.
-        if (!this.isOwner(participant)) {
+        if (!this.commandHandler.isOwner(participant)) {
             log.info('Ignoring message from non-owner address', { address: participant.slice(0, 8) + '...' });
             return;
         }
 
         // Emit feed event only for external (non-agent) messages
-        this.emitEvent(participant, content, 'inbound', fee);
+        this.responseFormatter.emitEvent(participant, content, 'inbound', fee);
 
         // ── Credit system: detect payments and convert to credits ─────────
-        // Any ALGO sent above minimum transaction fee gets converted to credits.
-        // The fee field from AlgoChat represents the total payment amount.
-        const MIN_TXFEE_MICRO = 1000; // Minimum Algorand txn fee
+        const MIN_TXFEE_MICRO = 1000;
         if (fee && fee > MIN_TXFEE_MICRO) {
-            const paymentMicro = fee - MIN_TXFEE_MICRO; // Subtract base fee
+            const paymentMicro = fee - MIN_TXFEE_MICRO;
             if (paymentMicro > 0) {
                 const creditsAdded = purchaseCredits(this.db, participant, paymentMicro);
                 if (creditsAdded > 0) {
@@ -875,10 +555,8 @@ export class AlgoChatBridge {
                         creditsAdded,
                         newBalance: balance.available,
                     });
-                    // Don't send a separate message for small purchases bundled with a chat message
-                    // Only notify for explicit large payments (> 0.1 ALGO)
                     if (paymentMicro >= 100_000) {
-                        this.sendResponse(participant,
+                        this.responseFormatter.sendResponse(participant,
                             `💰 +${creditsAdded} credits purchased! Balance: ${balance.available} credits`
                         );
                     }
@@ -890,7 +568,7 @@ export class AlgoChatBridge {
         const firstTimeCredits = maybeGrantFirstTimeCredits(this.db, participant);
         if (firstTimeCredits > 0) {
             const balance = getBalance(this.db, participant);
-            this.sendResponse(participant,
+            this.responseFormatter.sendResponse(participant,
                 `🎉 Welcome! You've received ${firstTimeCredits} free credits to get started.\n` +
                 `Balance: ${balance.available} credits\n` +
                 `Use /credits to check balance, send ALGO to buy more.`
@@ -898,7 +576,7 @@ export class AlgoChatBridge {
         }
 
         // Check for commands first
-        if (this.handleCommand(participant, content)) return;
+        if (this.commandHandler.handleCommand(participant, content)) return;
 
         // ── Credit check: ensure participant has credits before proceeding ──
         const creditCheck = canStartSession(this.db, participant);
@@ -907,7 +585,7 @@ export class AlgoChatBridge {
                 participant: participant.slice(0, 8) + '...',
                 credits: creditCheck.credits,
             });
-            this.sendResponse(participant,
+            this.responseFormatter.sendResponse(participant,
                 `⚠️ Insufficient credits (${creditCheck.credits} remaining).\n` +
                 `Send ALGO to purchase credits (1 ALGO = ${getCreditConfig(this.db).creditsPerAlgo} credits).\n` +
                 `Use /credits to check your balance.`
@@ -935,7 +613,7 @@ export class AlgoChatBridge {
         let conversation = getConversationByParticipant(this.db, participant);
 
         if (!conversation) {
-            const agentId = this.findAgentForNewConversation();
+            const agentId = this.discoveryService.findAgentForNewConversation();
             if (!agentId) {
                 log.info('No AlgoChat-enabled agent found, ignoring message');
                 return;
@@ -945,7 +623,7 @@ export class AlgoChatBridge {
             if (!agent) return;
 
             const session = createSession(this.db, {
-                projectId: agent.defaultProjectId ?? this.getDefaultProjectId(),
+                projectId: agent.defaultProjectId ?? this.discoveryService.getDefaultProjectId(),
                 agentId,
                 name: `AlgoChat: ${participant.slice(0, 8)}...`,
                 initialPrompt: content,
@@ -954,7 +632,7 @@ export class AlgoChatBridge {
 
             conversation = createConversation(this.db, participant, agentId, session.id);
 
-            this.subscribeForResponse(session.id, participant);
+            this.subscriptionManager.subscribeForResponse(session.id, participant);
 
             // Handle session start failure
             try {
@@ -964,7 +642,7 @@ export class AlgoChatBridge {
                     sessionId: session.id,
                     error: err instanceof Error ? err.message : String(err),
                 });
-                this.sendResponse(participant, `[Error: Failed to start agent session]`);
+                this.responseFormatter.sendResponse(participant, `[Error: Failed to start agent session]`);
             }
         } else {
             if (conversation.sessionId) {
@@ -973,13 +651,13 @@ export class AlgoChatBridge {
                     const { getSession } = await import('../db/sessions');
                     const session = getSession(this.db, conversation.sessionId);
                     if (session) {
-                        this.subscribeForResponse(session.id, participant);
+                        this.subscriptionManager.subscribeForResponse(session.id, participant);
                         this.processManager.resumeProcess(session, content);
                     }
                 }
             } else {
                 // Conversation exists but has no active session — create a new one
-                const agentId = conversation.agentId ?? this.findAgentForNewConversation();
+                const agentId = conversation.agentId ?? this.discoveryService.findAgentForNewConversation();
                 if (!agentId) {
                     log.info('No agent available for existing conversation, ignoring message');
                 } else {
@@ -987,7 +665,7 @@ export class AlgoChatBridge {
                     if (agent) {
                         const { updateConversationSession } = await import('../db/sessions');
                         const session = createSession(this.db, {
-                            projectId: agent.defaultProjectId ?? this.getDefaultProjectId(),
+                            projectId: agent.defaultProjectId ?? this.discoveryService.getDefaultProjectId(),
                             agentId,
                             name: `AlgoChat: ${participant.slice(0, 8)}...`,
                             initialPrompt: content,
@@ -996,7 +674,7 @@ export class AlgoChatBridge {
                         updateConversationSession(this.db, conversation.id, session.id);
                         conversation.sessionId = session.id;
 
-                        this.subscribeForResponse(session.id, participant);
+                        this.subscriptionManager.subscribeForResponse(session.id, participant);
                         try {
                             this.processManager.startProcess(session, content);
                         } catch (err) {
@@ -1004,7 +682,7 @@ export class AlgoChatBridge {
                                 sessionId: session.id,
                                 error: err instanceof Error ? err.message : String(err),
                             });
-                            this.sendResponse(participant, `[Error: Failed to start agent session]`);
+                            this.responseFormatter.sendResponse(participant, `[Error: Failed to start agent session]`);
                         }
                     }
                 }
@@ -1012,845 +690,5 @@ export class AlgoChatBridge {
         }
 
         updateConversationRound(this.db, conversation.id, confirmedRound);
-    }
-
-    private subscribeForResponse(sessionId: string, participant: string): void {
-        // Avoid duplicate subscriptions when multiple messages arrive for the same session
-        if (this.chainSubscriptions.has(sessionId)) return;
-        this.chainSubscriptions.add(sessionId);
-
-        // We only send the LAST text block from the last turn. Earlier text
-        // blocks are intermediate explanations (tool call reasoning, etc.)
-        // and would clutter the on-chain response.
-        let lastTextBlock = '';
-        let lastAssistantText = '';
-        let lastTurnResponse = '';
-        let sent = false;
-        const startedAt = Date.now();
-
-        const sendOnce = () => {
-            if (sent) return;
-            sent = true;
-
-            // Track completion milestone
-            const totalElapsed = Date.now() - startedAt;
-            trackProgress({
-                type: 'milestone',
-                action: 'response_completed',
-                timestamp: Date.now(),
-                details: `Total time: ${Math.round(totalElapsed / 1000)}s, tools: ${toolsUsed.size}, agents: ${agentsQueried.size}`
-            });
-
-            stopProgressTimer();
-            this.processManager.unsubscribe(sessionId, callback);
-            this.chainSubscriptions.delete(sessionId);
-            this.clearSubscriptionTimer(sessionId);
-
-            // Prefer streamed text block > last turn response > full assistant text
-            const finalText = (lastTextBlock.trim() || lastTurnResponse.trim() || lastAssistantText.trim());
-            if (finalText) {
-                this.sendResponse(participant, finalText);
-            }
-        };
-
-        const resetTimer = () => {
-            this.setSubscriptionTimer(sessionId, () => {
-                log.warn(`Subscription timeout for session ${sessionId}`);
-                sendOnce();
-            });
-        };
-
-        let statusEmitted = false;
-        let ackSent = false;
-        let agentQueryCount = 0;
-        let currentTextBlock = '';
-        let inTextBlock = false;
-
-        // Enhanced progress tracking
-        interface ProgressAction {
-            type: 'tool_use' | 'agent_query' | 'text_block' | 'milestone';
-            action: string;
-            timestamp: number;
-            details?: string;
-        }
-        const MAX_PROGRESS_HISTORY = 100;
-        const progressHistory: ProgressAction[] = [];
-        const trackProgress = (action: ProgressAction) => {
-            progressHistory.push(action);
-            // Sliding window to prevent unbounded memory growth in long sessions
-            if (progressHistory.length > MAX_PROGRESS_HISTORY) {
-                progressHistory.splice(0, progressHistory.length - MAX_PROGRESS_HISTORY);
-            }
-        };
-        let lastProgressUpdate = startedAt;
-        let toolsUsed: Set<string> = new Set();
-        let agentsQueried: Set<string> = new Set();
-        let progressTimer: ReturnType<typeof setInterval> | null = null;
-        let ackDelayTimer: ReturnType<typeof setTimeout> | null = null;
-
-        // How long to wait before sending the on-chain ack. If the response
-        // arrives within this window we skip the ack entirely.
-        const ACK_DELAY_MS = 10_000; // 10 seconds
-        const PROGRESS_INTERVAL_MS = 2 * 60 * 1000; // 2 minutes
-
-        const cancelAckDelay = () => {
-            if (ackDelayTimer) {
-                clearTimeout(ackDelayTimer);
-                ackDelayTimer = null;
-            }
-        };
-
-        // Generate progress summary from recent actions
-        const generateProgressSummary = (): string => {
-            const now = Date.now();
-            const elapsed = Math.round((now - startedAt) / 1000);
-            const recentActions = progressHistory.filter(a => a.timestamp > lastProgressUpdate);
-
-            let summary = `Still working (${elapsed}s elapsed)`;
-
-            // Add what we've been doing
-            const recentSummary: string[] = [];
-
-            if (recentActions.length > 0) {
-                const toolActions = recentActions.filter(a => a.type === 'tool_use');
-                const agentActions = recentActions.filter(a => a.type === 'agent_query');
-                const textActions = recentActions.filter(a => a.type === 'text_block');
-
-                if (toolActions.length > 0) {
-                    const uniqueTools = [...new Set(toolActions.map(a => a.action))];
-                    recentSummary.push(`used ${uniqueTools.join(', ')}`);
-                }
-
-                if (agentActions.length > 0) {
-                    const uniqueAgents = [...new Set(agentActions.map(a => a.action))];
-                    recentSummary.push(`queried ${uniqueAgents.join(', ')}`);
-                }
-
-                if (textActions.length > 0) {
-                    const lastText = textActions[textActions.length - 1];
-                    if (lastText.details && lastText.details.length > 0) {
-                        const preview = lastText.details.length > 60
-                            ? lastText.details.slice(0, 60) + '...'
-                            : lastText.details;
-                        recentSummary.push(`working on: ${preview}`);
-                    }
-                }
-            }
-
-            // Overall progress
-            if (toolsUsed.size > 0 || agentsQueried.size > 0) {
-                const progress: string[] = [];
-                if (toolsUsed.size > 0) {
-                    progress.push(`${toolsUsed.size} tool${toolsUsed.size > 1 ? 's' : ''}`);
-                }
-                if (agentsQueried.size > 0) {
-                    progress.push(`${agentsQueried.size} agent${agentsQueried.size > 1 ? 's' : ''}`);
-                }
-                summary += ` — used ${progress.join(' and ')}`;
-            }
-
-            if (recentSummary.length > 0) {
-                summary += ` — recently ${recentSummary.join(', ')}`;
-            }
-
-            lastProgressUpdate = now;
-            return summary;
-        };
-
-        // Send periodic on-chain progress updates so the user's AlgoChat
-        // client knows the agent is still working
-        const startProgressTimer = () => {
-            if (progressTimer) return;
-            progressTimer = setInterval(() => {
-                if (sent) { stopProgressTimer(); return; }
-                const msg = generateProgressSummary();
-                this.sendResponse(participant, `[Status] ${msg}`).catch((err) => {
-                    log.warn('Failed to send progress update', {
-                        participant,
-                        sessionId,
-                        error: err instanceof Error ? err.message : String(err),
-                    });
-                });
-                this.emitEvent(participant, msg, 'status');
-            }, PROGRESS_INTERVAL_MS);
-        };
-
-        const stopProgressTimer = () => {
-            if (progressTimer) {
-                clearInterval(progressTimer);
-                progressTimer = null;
-            }
-        };
-
-        // Actually send the on-chain ack and start progress timer
-        const sendAckNow = () => {
-            if (ackSent || sent) return;
-            ackSent = true;
-
-            // Track acknowledgment milestone
-            trackProgress({
-                type: 'milestone',
-                action: 'request_acknowledged',
-                timestamp: Date.now(),
-                details: 'Processing request'
-            });
-
-            this.sendResponse(participant, '[Status] Received your message — working on it now.').catch((err) => {
-                log.warn('Failed to send acknowledgment', {
-                    participant,
-                    sessionId,
-                    error: err instanceof Error ? err.message : String(err),
-                });
-            });
-            startProgressTimer();
-        };
-
-        const flushTextBlock = () => {
-            const text = currentTextBlock.trim();
-            if (text.length > 0) {
-                // Keep track of the latest text block — this overwrites
-                // previous ones so we only send the final one on-chain.
-                lastTextBlock = text;
-
-                // Track meaningful text for progress summaries
-                if (text.length > 50) { // Only track substantial text blocks
-                    trackProgress({
-                        type: 'text_block',
-                        action: 'reasoning',
-                        timestamp: Date.now(),
-                        details: text
-                    });
-                }
-
-                // Show the agent's intermediate text as a status update
-                // Truncate long blocks to a reasonable preview
-                const preview = text.length > 300
-                    ? text.slice(0, 300) + '...'
-                    : text;
-                this.emitEvent(participant, preview, 'status');
-            }
-            currentTextBlock = '';
-            inTextBlock = false;
-        };
-
-        const callback = (sid: string, event: ClaudeStreamEvent) => {
-            if (sid !== sessionId) return;
-
-            // On first assistant event, show a local status and schedule
-            // the on-chain ack after a delay (skip if we finish quickly)
-            if (event.type === 'assistant' && !statusEmitted) {
-                statusEmitted = true;
-
-                // Track processing milestone
-                trackProgress({
-                    type: 'milestone',
-                    action: 'processing_started',
-                    timestamp: Date.now(),
-                    details: 'Agent began processing'
-                });
-
-                this.emitEvent(participant, 'Agent is processing your message...', 'status');
-
-                if (!ackSent && !ackDelayTimer) {
-                    ackDelayTimer = setTimeout(sendAckNow, ACK_DELAY_MS);
-                }
-            }
-
-            // Forward named status events from tool handlers (e.g. "Querying CorvidLabs...")
-            if ((event as { type: string }).type === 'tool_status') {
-                const message = (event as unknown as { message: string }).message;
-                if (message) {
-                    this.emitEvent(participant, message, 'status');
-
-                    // Track status action for progress summaries
-                    trackProgress({
-                        type: 'milestone',
-                        action: 'status_update',
-                        timestamp: Date.now(),
-                        details: message
-                    });
-
-                    // Agent is calling other agents — this will take a while,
-                    // send the ack immediately
-                    if (!ackSent) {
-                        cancelAckDelay();
-                        sendAckNow();
-                    }
-                    resetTimer();
-                }
-                return;
-            }
-
-            // Track text content blocks — stream agent's intermediate text to the feed
-            if (event.type === 'content_block_start') {
-                const block = event.content_block;
-                if (block?.type === 'text') {
-                    inTextBlock = true;
-                    currentTextBlock = '';
-                } else if (block?.type === 'tool_use') {
-                    // Flush any pending text before tool use starts
-                    if (inTextBlock) flushTextBlock();
-                    const toolName = (block as unknown as { name?: string }).name;
-
-                    if (toolName) {
-                        // Track tool usage for progress summaries
-                        toolsUsed.add(toolName);
-                        trackProgress({
-                            type: 'tool_use',
-                            action: toolName,
-                            timestamp: Date.now()
-                        });
-
-                        if (toolName === 'corvid_send_message') {
-                            agentQueryCount++;
-                            // Try to extract agent name from tool input
-                            const toolInput = (block as unknown as { input?: { to_agent?: string } }).input;
-                            if (toolInput?.to_agent) {
-                                agentsQueried.add(toolInput.to_agent);
-                                trackProgress({
-                                    type: 'agent_query',
-                                    action: toolInput.to_agent,
-                                    timestamp: Date.now()
-                                });
-                            }
-                            // Agent-to-agent call means longer processing — send ack now
-                            if (!ackSent) {
-                                cancelAckDelay();
-                                sendAckNow();
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Accumulate streaming text deltas
-            if (event.type === 'content_block_delta' && event.delta?.text && inTextBlock) {
-                currentTextBlock += event.delta.text;
-                resetTimer();
-            }
-
-            // Text block finished — flush it as a status update
-            if (event.type === 'content_block_stop' && inTextBlock) {
-                flushTextBlock();
-            }
-
-            // Capture assistant message content as fallback in case content_block
-            // streaming events don't fire (e.g. non-streaming SDK responses)
-            if (event.type === 'assistant' && event.message?.content) {
-                const text = extractContentText(event.message.content);
-                if (text.trim()) {
-                    lastAssistantText = text;
-                }
-                resetTimer();
-            } else if (event.type === 'assistant') {
-                resetTimer();
-            }
-
-            // Each 'result' marks end of a turn — save last text block and reset
-            if (event.type === 'result') {
-                if (inTextBlock) flushTextBlock();
-
-                // Track turn completion milestone
-                trackProgress({
-                    type: 'milestone',
-                    action: 'turn_completed',
-                    timestamp: Date.now(),
-                    details: `Completed turn with ${agentQueryCount} agent queries`
-                });
-
-                // Only show synthesizing status if we've been working long enough
-                const elapsed = Date.now() - startedAt;
-                if (agentQueryCount > 0 && elapsed > ACK_DELAY_MS) {
-                    const synthesizingMsg = `Synthesizing response from ${agentQueryCount} agent${agentQueryCount > 1 ? 's' : ''}...`;
-
-                    // Track synthesis milestone
-                    trackProgress({
-                        type: 'milestone',
-                        action: 'synthesis_started',
-                        timestamp: Date.now(),
-                        details: synthesizingMsg
-                    });
-
-                    this.emitEvent(participant, synthesizingMsg, 'status');
-                }
-                // Prefer the last streamed text block; fall back to full assistant text
-                if (lastTextBlock.trim()) {
-                    lastTurnResponse = lastTextBlock;
-                } else if (lastAssistantText.trim()) {
-                    lastTurnResponse = lastAssistantText;
-                }
-                lastTextBlock = '';
-                lastAssistantText = '';
-                resetTimer(); // Turn completed — reset timeout
-            }
-
-            // Send only the last turn's response when the session fully exits
-            if (event.type === 'session_exited') {
-                if (inTextBlock) flushTextBlock();
-                cancelAckDelay();
-                stopProgressTimer();
-                sendOnce();
-            }
-        };
-
-        this.processManager.subscribe(sessionId, callback);
-        resetTimer();
-    }
-
-    private subscribeForLocalResponse(sessionId: string, sendFn: LocalChatSendFn): void {
-        // Store the sendFn so it can be updated if the WS connection changes
-        this.localSendFns.set(sessionId, sendFn);
-
-        // Check if already subscribed (avoid duplicate subscriptions on subsequent messages)
-        if (this.localSubscriptions.has(sessionId)) return;
-
-        let responseBuffer = '';
-        let isThinking = false;
-
-        const callback = (sid: string, event: ClaudeStreamEvent) => {
-            if (sid !== sessionId) return;
-
-            // Always use the latest sendFn and eventFn
-            const currentSendFn = this.localSendFns.get(sessionId);
-            if (!currentSendFn) return;
-            const currentEventFn = this.localEventFns.get(sessionId);
-
-            log.debug(`Local response event`, { sessionId, type: event.type, subtype: event.subtype });
-
-            // Emit thinking events
-            if (event.type === 'assistant' && !isThinking) {
-                isThinking = true;
-                currentEventFn?.({ type: 'thinking', active: true });
-            }
-
-            // Emit streaming chunks for content_block_delta
-            if (event.type === 'content_block_delta' && event.delta?.text) {
-                currentEventFn?.({ type: 'stream', chunk: event.delta.text, done: false });
-            }
-
-            // Emit tool_use events
-            if (event.type === 'content_block_start' && event.content_block?.type === 'tool_use') {
-                const toolName = (event.content_block as unknown as { name?: string }).name ?? 'unknown';
-                const input = JSON.stringify((event.content_block as unknown as { input?: unknown }).input ?? {});
-                currentEventFn?.({ type: 'tool_use', toolName, input });
-            }
-
-            if (event.type === 'assistant' && event.message?.content) {
-                const text = extractContentText(event.message.content);
-                log.debug(`Assistant content chunk`, { text: text.slice(0, 80) });
-                responseBuffer += text;
-            }
-
-            // Turn completed — send accumulated response and reset buffer for next turn
-            if (event.type === 'result') {
-                log.debug(`Turn completed`, { bufferLength: responseBuffer.length });
-                isThinking = false;
-                currentEventFn?.({ type: 'thinking', active: false });
-                currentEventFn?.({ type: 'stream', chunk: '', done: true });
-
-                if (responseBuffer.trim()) {
-                    log.debug(`Sending outbound response`, { text: responseBuffer.trim().slice(0, 80) });
-                    currentSendFn('local', responseBuffer.trim(), 'outbound');
-                    currentEventFn?.({ type: 'message', content: responseBuffer.trim(), direction: 'outbound' });
-                }
-                responseBuffer = '';
-            }
-
-            // Session exited — clean up subscription
-            if (event.type === 'session_exited') {
-                log.debug('Session exited, cleaning up subscription');
-                this.processManager.unsubscribe(sessionId, callback);
-                this.localSubscriptions.delete(sessionId);
-                this.localSendFns.delete(sessionId);
-                this.localEventFns.delete(sessionId);
-                this.clearSubscriptionTimer(sessionId);
-
-                isThinking = false;
-                currentEventFn?.({ type: 'thinking', active: false });
-
-                // Send any remaining buffered text
-                if (responseBuffer.trim()) {
-                    currentSendFn('local', responseBuffer.trim(), 'outbound');
-                    currentEventFn?.({ type: 'message', content: responseBuffer.trim(), direction: 'outbound' });
-                }
-            }
-        };
-
-        this.localSubscriptions.set(sessionId, callback);
-        this.processManager.subscribe(sessionId, callback);
-        this.setSubscriptionTimer(sessionId, () => {
-            log.warn(`Local subscription timeout for session ${sessionId}`);
-            this.processManager.unsubscribe(sessionId, callback);
-            this.localSubscriptions.delete(sessionId);
-            const currentSendFn = this.localSendFns.get(sessionId);
-            this.localSendFns.delete(sessionId);
-            this.localEventFns.delete(sessionId);
-            if (responseBuffer.trim() && currentSendFn) {
-                currentSendFn('local', responseBuffer.trim(), 'outbound');
-            }
-        });
-    }
-
-    private setSubscriptionTimer(sessionId: string, onTimeout: () => void): void {
-        // Clear any existing timer for this session
-        this.clearSubscriptionTimer(sessionId);
-        const timer = setTimeout(onTimeout, SUBSCRIPTION_TIMEOUT_MS);
-        this.subscriptionTimers.set(sessionId, timer);
-    }
-
-    private clearSubscriptionTimer(sessionId: string): void {
-        const timer = this.subscriptionTimers.get(sessionId);
-        if (timer) {
-            clearTimeout(timer);
-            this.subscriptionTimers.delete(sessionId);
-        }
-    }
-
-    /** Split content into byte-limited chunks for PSK sends, breaking at newlines when possible. */
-    private splitPskContent(content: string, maxBytes: number): string[] {
-        const encoder = new TextEncoder();
-        if (encoder.encode(content).byteLength <= maxBytes) {
-            return [content];
-        }
-
-        const chunks: string[] = [];
-        let remaining = content;
-
-        while (remaining.length > 0) {
-            if (encoder.encode(remaining).byteLength <= maxBytes) {
-                chunks.push(remaining);
-                break;
-            }
-
-            // Binary search for the max character count that fits in maxBytes
-            let low = 0;
-            let high = remaining.length;
-            while (low < high) {
-                const mid = Math.floor((low + high + 1) / 2);
-                if (encoder.encode(remaining.slice(0, mid)).byteLength <= maxBytes) {
-                    low = mid;
-                } else {
-                    high = mid - 1;
-                }
-            }
-
-            // Try to break at a newline within the last 20% for readability
-            let cut = low;
-            const searchStart = Math.floor(low * 0.8);
-            const lastNewline = remaining.lastIndexOf('\n', low);
-            if (lastNewline >= searchStart) {
-                cut = lastNewline + 1;
-            }
-
-            chunks.push(remaining.slice(0, cut));
-            remaining = remaining.slice(cut);
-        }
-
-        return chunks;
-    }
-
-    private async sendResponse(participant: string, content: string): Promise<void> {
-        // Check daily ALGO spending limit (estimate min fee of 1000 microAlgos per txn)
-        try {
-            checkAlgoLimit(this.db, 1000);
-        } catch (err) {
-            const conversation = getConversationByParticipant(this.db, participant);
-            log.warn(`On-chain response blocked by spending limit — dead letter`, {
-                participant,
-                conversationId: conversation?.id ?? null,
-                sessionId: conversation?.sessionId ?? null,
-                contentLength: content.length,
-                contentPreview: content.slice(0, 200),
-                error: err instanceof Error ? err.message : String(err),
-            });
-            return;
-        }
-
-        try {
-            // Route PSK contacts through the PSK manager
-            if (this.pskManager && participant === this.pskManager.contactAddress) {
-                // PSK has an 878-byte payload limit per transaction.
-                // Split oversized messages into sequential sends with a delay
-                // between each so they land in different blocks (preserving order).
-                const PSK_MAX_BYTES = 800;
-                const PSK_INTER_CHUNK_DELAY_MS = 4500;
-                const chunks = this.splitPskContent(content, PSK_MAX_BYTES);
-                for (let i = 0; i < chunks.length; i++) {
-                    if (i > 0) {
-                        await new Promise((r) => setTimeout(r, PSK_INTER_CHUNK_DELAY_MS));
-                    }
-                    await this.pskManager.sendMessage(chunks[i]);
-                }
-                log.info(`Sent PSK response to ${participant}`, {
-                    content: content.slice(0, 100),
-                    chunks: chunks.length,
-                });
-                this.emitEvent(participant, content, 'outbound');
-                return;
-            }
-
-            // Try to use per-agent wallet if available
-            let senderAccount = this.service.chatAccount;
-            if (this.agentWalletService) {
-                const conversation = getConversationByParticipant(this.db, participant);
-                if (conversation?.agentId) {
-                    const agentAccount = await this.agentWalletService.getAgentChatAccount(conversation.agentId);
-                    if (agentAccount) {
-                        senderAccount = agentAccount.account;
-                        log.debug(`Using agent wallet ${agentAccount.address} for response`);
-                    }
-                }
-            }
-
-            const pubKey = await this.getPublicKey(participant);
-
-            // Use group transactions for all recipients. External AlgoChat
-            // clients that support [GRP:] reassembly will show the full message;
-            // for short messages that fit in a single txn, sendGroupMessage
-            // automatically falls back to a standard single send.
-            try {
-                const { sendGroupMessage } = await import('./group-sender');
-                const groupResult = await sendGroupMessage(
-                    this.service,
-                    senderAccount,
-                    participant,
-                    pubKey,
-                    content,
-                );
-
-                log.info(`Sent response to ${participant}`, { content: content.slice(0, 100), fee: groupResult.fee, txids: groupResult.txids.length });
-                if (groupResult.fee) {
-                    recordAlgoSpend(this.db, groupResult.fee);
-                    const conv = getConversationByParticipant(this.db, participant);
-                    if (conv?.sessionId) updateSessionAlgoSpent(this.db, conv.sessionId, groupResult.fee);
-                }
-                this.emitEvent(participant, content, 'outbound', groupResult.fee);
-                return;
-            } catch (groupErr) {
-                log.warn('Group send failed, falling back to single txn', {
-                    error: groupErr instanceof Error ? groupErr.message : String(groupErr),
-                });
-            }
-
-            // Fallback: single transaction (truncates if needed)
-            let sendContent = content;
-            const encoded = new TextEncoder().encode(content);
-            if (encoded.byteLength > 850) {
-                sendContent = new TextDecoder().decode(encoded.slice(0, 840)) + '...';
-            }
-
-            const result = await this.service.algorandService.sendMessage(
-                senderAccount,
-                participant,
-                pubKey,
-                sendContent,
-            );
-
-            const fee = (result as unknown as { fee?: number }).fee;
-            log.info(`Sent response to ${participant}`, { content: content.slice(0, 100), fee });
-            if (fee) {
-                recordAlgoSpend(this.db, fee);
-                const conv = getConversationByParticipant(this.db, participant);
-                if (conv?.sessionId) updateSessionAlgoSpent(this.db, conv.sessionId, fee);
-            }
-            this.emitEvent(participant, content, 'outbound', fee);
-        } catch (err) {
-            // Dead-letter logging: capture full context for failed message sends
-            // so they can be investigated and potentially retried.
-            const conversation = getConversationByParticipant(this.db, participant);
-            log.error('Failed to send response — dead letter', {
-                participant,
-                conversationId: conversation?.id ?? null,
-                sessionId: conversation?.sessionId ?? null,
-                agentId: conversation?.agentId ?? null,
-                contentLength: content.length,
-                contentPreview: content.slice(0, 200),
-                error: err instanceof Error ? err.message : String(err),
-                stack: err instanceof Error ? err.stack : undefined,
-            });
-        }
-    }
-
-    private async getPublicKey(address: string): Promise<Uint8Array> {
-        const cached = this.publicKeyCache.get(address);
-        if (cached && (Date.now() - cached.cachedAt) < PUBLIC_KEY_CACHE_TTL_MS) {
-            return cached.key;
-        }
-
-        const pubKey = await this.service.algorandService.discoverPublicKey(address);
-        this.publicKeyCache.set(address, { key: pubKey, cachedAt: Date.now() });
-        return pubKey;
-    }
-
-    private findAgentForNewConversation(): string | null {
-        if (this.config.defaultAgentId) {
-            return this.config.defaultAgentId;
-        }
-
-        const agents = getAlgochatEnabledAgents(this.db);
-        const autoAgent = agents.find((a) => a.algochatAuto);
-        return autoAgent?.id ?? agents[0]?.id ?? null;
-    }
-
-    private getDefaultProjectId(): string {
-        const { listProjects, createProject } = require('../db/projects');
-        const projects = listProjects(this.db);
-        if (projects.length > 0) return projects[0].id;
-
-        const project = createProject(this.db, {
-            name: 'AlgoChat Default',
-            workingDir: process.cwd(),
-        });
-        return project.id;
-    }
-
-    /** Cache of agent wallet addresses, refreshed lazily. */
-    private cachedAgentWallets: Set<string> | null = null;
-    private cachedAgentWalletsAt = 0;
-
-    private getAgentWalletAddresses(): Set<string> {
-        const now = Date.now();
-        // Refresh cache every 60s
-        if (this.cachedAgentWallets && now - this.cachedAgentWalletsAt < 60_000) {
-            return this.cachedAgentWallets;
-        }
-        const agents = listAgents(this.db);
-        const addrs = new Set<string>();
-        for (const a of agents) {
-            if (a.walletAddress) addrs.add(a.walletAddress);
-        }
-        // Also include the main chat account address
-        addrs.add(this.service.chatAccount.address);
-        this.cachedAgentWallets = addrs;
-        this.cachedAgentWalletsAt = now;
-        return addrs;
-    }
-
-    private emitEvent(
-        participant: string,
-        content: string,
-        direction: 'inbound' | 'outbound' | 'status',
-        fee?: number,
-    ): void {
-        // Persist to DB so messages survive page refresh
-        try {
-            saveAlgoChatMessage(this.db, { participant, content, direction, fee });
-        } catch (err) {
-            log.warn('Failed to persist algochat message', { error: err instanceof Error ? err.message : String(err) });
-        }
-
-        for (const cb of this.eventCallbacks) {
-            try {
-                cb(participant, content, direction, fee);
-            } catch (err) {
-                log.error('Event callback threw', { error: err instanceof Error ? err.message : String(err) });
-            }
-        }
-    }
-
-    /**
-     * While approval requests are pending, poll on-chain at a faster rate
-     * (5s) so the user's reply is picked up quickly.
-     */
-    private startFastPolling(): void {
-        if (this.fastPollTimer) return;
-
-        const FAST_POLL_MS = 5000;
-        this.fastPollTimer = setInterval(() => {
-            // If no more pending approvals, stop fast polling
-            if (!this.approvalManager?.hasPendingRequests()) {
-                this.stopFastPolling();
-                return;
-            }
-
-            // Trigger a manual sync
-            this.service.syncManager.sync().catch((err) => {
-                log.warn('Fast-poll sync error', { error: err instanceof Error ? err.message : String(err) });
-            });
-        }, FAST_POLL_MS);
-
-        log.debug('Started fast-polling for approval responses');
-    }
-
-    private stopFastPolling(): void {
-        if (this.fastPollTimer) {
-            clearInterval(this.fastPollTimer);
-            this.fastPollTimer = null;
-            log.debug('Stopped fast-polling');
-        }
-    }
-
-    /**
-     * Seed the SyncManager with known conversation participants from the DB
-     * so that fetchAllConversations has something to iterate over.
-     */
-    private seedConversations(): void {
-        const conversations = listConversations(this.db);
-        for (const conv of conversations) {
-            const syncConv = this.service.syncManager.getOrCreateConversation(conv.participantAddr);
-            if (conv.lastRound > 0) {
-                // Use lastRound + 1 because minRound is inclusive and we already
-                // processed the message at lastRound in a previous run
-                syncConv.setLastFetchedRound(conv.lastRound + 1);
-            }
-        }
-        if (conversations.length > 0) {
-            log.info(`Seeded ${conversations.length} conversation(s) from DB`);
-        }
-    }
-
-    /**
-     * Periodically discover new senders by querying the indexer for
-     * incoming transactions from addresses not yet in the SyncManager.
-     */
-    private startDiscoveryPolling(): void {
-        if (this.discoveryTimer) return;
-
-        // Run immediately then on the sync interval
-        this.discoverNewSenders().catch((err) => {
-            log.warn('Discovery error', { error: err instanceof Error ? err.message : String(err) });
-        });
-
-        this.discoveryTimer = setInterval(() => {
-            this.discoverNewSenders().catch((err) => {
-                log.warn('Discovery error', { error: err instanceof Error ? err.message : String(err) });
-            });
-        }, this.config.syncInterval);
-    }
-
-    private stopDiscoveryPolling(): void {
-        if (this.discoveryTimer) {
-            clearInterval(this.discoveryTimer);
-            this.discoveryTimer = null;
-        }
-    }
-
-    private async discoverNewSenders(): Promise<void> {
-        if (!this.service.indexerClient) return;
-
-        const myAddr = this.service.chatAccount.address;
-        const response = await this.service.indexerClient
-            .searchForTransactions()
-            .address(myAddr)
-            .addressRole('receiver')
-            .limit(50)
-            .do();
-
-        const knownParticipants = new Set(
-            this.service.syncManager.getConversations().map((c) => c.participant),
-        );
-
-        const newSenders = new Set<string>();
-        for (const tx of (response as { transactions?: Array<{ sender: string; note?: string }> }).transactions ?? []) {
-            if (tx.sender !== myAddr && tx.note && !knownParticipants.has(tx.sender)) {
-                if (!this.isOwner(tx.sender)) continue;
-                newSenders.add(tx.sender);
-            }
-        }
-
-        for (const sender of newSenders) {
-            log.info(`Discovered new sender`, { address: sender });
-            this.service.syncManager.getOrCreateConversation(sender);
-        }
     }
 }
